Now seems to be somewhat working after removing the use of the prod function (i.e relying explicity on * multiplication). Generally the point of this update is to be able to compose PQCs 'from the ground up' i.e by defining lots of circuits that one can piece together to form larger structures. Ideally the class system and inheritance will allow these groups of circuits to display desirable proprties i.e parameterised composite gates (for example ZZ) being derivable, and therefore trainable.
To this end a generic placeholder gate class has been defined that only really contains the multiplication behaviour of these subclasses,
which is to fall back onto normal qutip multiplication behaviour. Then two subclasses have been defined from that, PRot and EntGate,
for Parameterised Rotations and Entangling gates specifically. Both types of gates have the number of qubits in the system total
(I know this is ugly but no real way around it afaik) and the qubits they act on (1 qubit for single-qubit gates, 2 for entangling gates).

PRot gates have an angle they rotate by (initialised at 0, and can be set later when circuit initialised) and an operation they perform.
Sub-children of PRot (sub-sub-children of Gate) like R_x, R_y, ..., overwrtie the operation with their appropriate qutip function.
It can be seen that this structure allows for easy Hadamard gate definition - just a constant pi/4 R_y gate.

Entangling gates have the qubits they act on and the total qubit number as arguments. Again like the rotation gates, sub-children have
their specific operation re-defined so as to act properly. Using these sub-children like CNOT, CPHASE etc can then easily define
entangling blocks like the CHAIN or ALL-TO-ALL connections that are parameterised by the entanglers they use and the number of qubits in the
system.

The main benefit of this is hopefully that arbritrary circuit structures are much easier to create, but also to represent - a PQC is now simply
an ordered list of gate operations that ideally match the literature notation. Work still needs to be done on implementing gradient stuff
with this new methodology (and then implementing training!!). In terms of testing - there's 2 new tests on tests.py that showcase this
formalism in action, the idle circuit and Circuit A (a Hadamard and R_x), with the results matching the expected values.

Documentation will arrive soon.